// src/types/schemas/todoList.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface TodoList {
	techSetup: {
		name: string;
		items: {
			name: string;
			completed: boolean;
			[k: string]: unknown;
		}[];
	}[];
}